+++ wear/src/main/java/info/nightscout/androidaps/watchfaces/CircleWatchface.java
@@ -73,7 +73,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
 
 
     private int batteryLevel = 0;
-    private double datetime = 0;
+    private long datetime = 0;
     private String direction = "";
     private String delta = "";
     private String avgDelta = "";
@@ -433,11 +433,11 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
     }
 
 
-    private synchronized double getDatetime() {
+    private synchronized long getDatetime() {
         return datetime;
     }
 
-    private synchronized void setDatetime(double datetime) {
+    private synchronized void setDatetime(long datetime) {
         this.datetime = datetime;
     }
 
@@ -534,7 +534,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
                 Log.d("CircleWatchface", "sgv string : " + getSgvString());
                 setDelta(dataMap.getString("delta"));
                 setAvgDelta(dataMap.getString("avgDelta"));
-                setDatetime(dataMap.getDouble("timestamp"));
+                setDatetime(dataMap.getLong("timestamp"));
                 addToWatchSet(dataMap);
 
 
@@ -576,7 +576,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
             double sgv = dataMap.getDouble("sgvDouble");
             double high = dataMap.getDouble("high");
             double low = dataMap.getDouble("low");
-            double timestamp = dataMap.getDouble("timestamp");
+            long timestamp = dataMap.getLong("timestamp");
             bgDataList.add(new BgWatchData(sgv, high, low, timestamp));
         } else if (!sharedPrefs.getBoolean("animation", false)) {
             // don't load history at once if animations are set (less resource consumption)
@@ -586,7 +586,7 @@ public class CircleWatchface extends WatchFace implements SharedPreferences.OnSh
                 double sgv = entry.getDouble("sgvDouble");
                 double high = entry.getDouble("high");
                 double low = entry.getDouble("low");
-                double timestamp = entry.getDouble("timestamp");
+                long timestamp = entry.getLong("timestamp");
                 bgDataList.add(new BgWatchData(sgv, high, low, timestamp));
             }
         } else
